rename auto_login plug to token_login
